… models/testdata/line-protocol.txt