### Overview

The behavior of the line-protocol parser remains unchanged. This PR moves the functions to a struct in order to simplify the tracking of limits whilst parsing a request.

Limits may be specified using the `ParsePointsWithOptions` API and the associated `WithParser*` functional options, all of which are documented.

It should be noted that the behavior of `ParsePointsWithPrecision` remains unchanged.

`ParsePointsWithPrecisionV1` was no longer used and removed to simplify code.

### Limits

The following limits are supported by the parser.

* `WithParserMaxBytes` specifies the maximum number of bytes that may be allocated when processing a single request. The parser accounts for the following allocations:

    * The size of the `[]models.Point` slice (16 bytes per element on 64-bit platforms)
    * The size and number of `*models.point` instances, which conform to the `models.Point` interface
    * The number of bytes allocated to store the transformed series key (V1 to V2)

* `WithParserMaxLines` specifies the maximum number of lines that may be parsed when processing a single request.

* `WithParserMaxValues` specifies the maximum number of values that may be parsed when processing a single request.

Any combination of these may be specified and if any limit is exceeded, parsing will fail with an appropriate error. Those which are omitted are considered unlimited.

### Stats

* `WithParserStats` specifies that `s` will contain statistics about the parsed request. This could be used to report the number of bytes required for parsing individual write requests.

### Benchmarks

Performance was not measurably impacted by these changes. In fact, performance has slightly improved.

```
↳ benchstat benchmarks/baseline.txt benchmarks/after.txt
name                                               old time/op    new time/op    delta
ParsePointsWithPrecision/gateway-toolsus1.txt-16      810µs ± 2%     781µs ± 1%  -3.58%  (p=0.000 n=10+9)
ParsePointsWithPrecision/gateway-us-west-2.txt-16     234ms ± 2%     223ms ± 2%  -4.65%  (p=0.000 n=10+10)

name                                               old speed      new speed      delta
ParsePointsWithPrecision/gateway-toolsus1.txt-16    144MB/s ± 2%   149MB/s ± 1%  +3.71%  (p=0.000 n=10+9)
ParsePointsWithPrecision/gateway-us-west-2.txt-16   154MB/s ± 2%   162MB/s ± 2%  +4.88%  (p=0.000 n=10+10)

name                                               old alloc/op   new alloc/op   delta
ParsePointsWithPrecision/gateway-toolsus1.txt-16      281kB ± 0%     281kB ± 0%  +0.07%  (p=0.000 n=8+9)
ParsePointsWithPrecision/gateway-us-west-2.txt-16    79.3MB ± 0%    79.3MB ± 0%  +0.00%  (p=0.000 n=9+9)

name                                               old allocs/op  new allocs/op  delta
ParsePointsWithPrecision/gateway-toolsus1.txt-16      1.17k ± 0%     1.17k ± 0%  +0.26%  (p=0.000 n=10+10)
ParsePointsWithPrecision/gateway-us-west-2.txt-16      311k ± 0%      311k ± 0%  +0.00%  (p=0.000 n=10+9)
```

### Usage via `/api/v2/write`

A subsequent PR will update the `handleWrite` function (`/api/v2/write`) to use the new `ParsePointsWithOptions` API  and will expose configurable limits.