@stuartcarnie 2020-01-10
Sorry gents, I'll remove the large file. git itself gzips, but I agree, it is unnecessary to have a single large file. I'll just repeat the small one, as it won't make any difference to the tests.